+++ apps/backend/src/modules/users/libs/types/types.ts
@@ -1,8 +1,11 @@
 export { type UserCredentials } from "./user-credentials.type.js";
+export { type UserDetails } from "./user-details.type.js";
 export {
 	type AuthResponseDto,
 	type UserGetAllResponseDto,
 	type UserResponseDto,
 	type UserSignInRequestDto,
 	type UserSignUpRequestDto,
+	type UserUpdateResponseDto,
+	type UserWithDetailsDto,
 } from "@meetlytic/shared";