+++ apps/frontend/src/libs/enums/app-route.enum.ts
@@ -3,6 +3,7 @@ const AppRoute = {
 	FEATURES: "#features",
 	MEETINGS: "/meetings",
 	MEETINGS_$ID: "/meetings/:id",
+	PROFILE: "/profile",
 	ROOT: "/",
 	SIGN_IN: "/sign-in",
 	SIGN_UP: "/sign-up",