+++ apps/frontend/src/libs/enums/avatar-size.enum.ts
@@ -1,5 +1,6 @@
 const AvatarSize = {
 	LARGE: "avatar-large",
+	MOBILE: "avatar-mobile",
 	SMALL: "avatar-small",
 } as const;