+++ apps/frontend/src/libs/types/icon-name.type.ts
@@ -1,5 +1,7 @@
 type IconName =
 	| "arrowRight"
+	| "burgerMenu"
+	| "closeIcon"
 	| "hidePassword"
 	| "logout"
 	| "meetings"