+++ apps/frontend/src/libs/types/types.ts
@@ -1,6 +1,4 @@
 export { type AsyncThunkConfig } from "./async-thunk-config.type.js";
-export { type Size } from "./button-size.type.js";
-export { type Variant } from "./button-variant.type.js";
 export { type IconName } from "./icon-name.type.js";
 export { type InputPassword } from "./input-password.type.js";
 export { type NavigationItem } from "./navigation-item.type.js";
@@ -8,6 +6,7 @@ export {
 	type ServerErrorDetail,
 	type ServerErrorResponse,
 	type UserResponseDto,
+	type UserWithDetailsDto,
 	type ValidationSchema,
 	type ValueOf,
 } from "@meetlytic/shared";