+++ apps/frontend/src/modules/users/libs/types/types.ts
@@ -2,4 +2,6 @@ export {
 	type UserResponseDto,
 	type UserSignInRequestDto,
 	type UserSignUpRequestDto,
+	type UserUpdateResponseDto,
+	type UserWithDetailsDto,
 } from "@meetlytic/shared";